PAYER/TENANTS/serializers.py
from rest_framework import serializers
from .models import Tenant, PaymentTransaction

class TenantSerializer(serializers.ModelSerializer):
    class Meta:
        model = Tenant
        fields = '__all__'

class PaymentTransactionSerializer(serializers.ModelSerializer):
    class Meta:
        model = PaymentTransaction
        fields = '__all__'