PAYER/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==3.2.23
django-cors-headers==4.1.0
djangorestframework==3.14.0
idna==3.4
mysqlclient==1.4.2.post1
Pillow==9.5.0
python-decouple==3.8
python-dotenv==0.21.1
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.7
gunicorn==21.2.0
dj-database-url==0.5.0
psycopg2==2.9.9
django-heroku==0.3.1